refactor(sidebar): Rename 'xc-sidebar' to 'sidebar'
