+++ packages/grid/stories/grid.stories.tsx
@@ -13,7 +13,7 @@ import {
 } from './data';
 import * as countries from './countries';
 import { gridThemeNamespace } from '../src/theme';
-import Sidebar from '../../xc-sidebar/src';
+import Sidebar from '../../sidebar/src';
 
 
 const list: any = (n: number) => (
