+++ packages/sidebar/package.json
