+++ packages/sidebar/src/interfaces.ts
