+++ packages/sidebar/src/styled/Sidebar.tsx
