+++ packages/sidebar/src/utils/getStyles.ts
