+++ packages/sidebar/tsconfig.json
